requirements.txt
flet
pymupdf
pypdf
fpdf2
